+++ backend/budget/admin.py
@@ -1,4 +1,12 @@
 from django.contrib import admin
 from budget.models import Budget
 
-admin.site.register(Budget)
+
+class CustomBudgetAdmin(admin.ModelAdmin):
+    list_display = ("name", "total_budget")
+    list_filter = ("name", "total_budget")
+    search_fields = ("name",)
+    ordering = ("name",)
+
+
+admin.site.register(Budget, CustomBudgetAdmin)