+++ backend/budget/views/common/requests/budget.py
@@ -7,16 +7,16 @@ from common.views.common.requests.base import BaseRequest, Headers
 
 
 @dataclass(frozen=True)  # TODO: pydantic validation
-class CredentialRequest(BaseRequest):
-    email: str
-    password: str
+class BudgetRequest(BaseRequest):
+    name: str
+    total_budget: int
 
     @staticmethod
-    def _init(request: Request) -> 'CredentialRequest':
+    def _init(request: Request) -> 'BudgetRequest':
         request_data = dict(request.data)
-        return CredentialRequest(
-            email=str(request_data.get('email')),
-            password=str(request_data.get('password')),
+        return BudgetRequest(
+            name=str(request_data.get('name')),
+            total_budget=int(request_data.get('total_budget')),
             user=str(request.user),
             content_type=str(request.content_type),
             auth=request.auth,
@@ -24,11 +24,11 @@ class CredentialRequest(BaseRequest):
         )
 
     @staticmethod
-    def init(request_data: Request) -> 'CredentialRequest' | Response:
+    def init(request_data: Request) -> 'BudgetRequest' | Response:
         try:
-            return CredentialRequest._init(request_data)
+            return BudgetRequest._init(request_data)
         except TypeError:
             return Response(
-                {"text": "Request data is not valid. Please provide 'email' and 'password' fields."},
+                {"text": "Request data is not valid. Please provide 'name' and 'total_budget' fields."},
                 status=status.HTTP_400_BAD_REQUEST,
             )