+++ backend/common/views/rest.py
@@ -1,13 +1,7 @@
-from django.views import generic
-
 from rest_framework import viewsets, status
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework.permissions import AllowAny
-
-
-class IndexView(generic.TemplateView):
-    template_name = 'common/index.html'
 
 
 class RestViewSet(viewsets.ViewSet):