+++ docs/CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [[0.4.3]] - 2022-03-24
+
+###### _([diff: 0.4.1-0.4.2])_
+
+### Added
+
+- Filtering & Display for `Budget` in Django Admin View
+- Adding new Budget via API Call (integrated)
+
+### To-do:
+
+- Handle new budget API call response on frontend.
+
 ## [[0.4.2]] - 2022-03-24
 
 ###### _([diff: 0.4.1-0.4.2])_
@@ -227,6 +240,7 @@ Connected Frontend & Backend via webpack_loader.
     -   I know I should implement them right away and have all of that sort of things already done, but I _really_ want to start some dev coding 🐈
     -   ... still got to setup frontend-backend, so I'll prioritize that at this moment.
 
+[diff: 0.4.2-0.4.3]: https://github.com/Luzkan/FamilyBudget/compare/0.4.2...0.4.3
 [diff: 0.4.1-0.4.2]: https://github.com/Luzkan/FamilyBudget/compare/0.4.1...0.4.2
 [diff: 0.4.0-0.4.1]: https://github.com/Luzkan/FamilyBudget/compare/0.4.0...0.4.1
 [diff: 0.4.0-alpha-0.4.0]: https://github.com/Luzkan/FamilyBudget/compare/0.4.0-alpha...0.4.0
@@ -242,6 +256,7 @@ Connected Frontend & Backend via webpack_loader.
 [diff: 0.1.1-0.2.0]: https://github.com/Luzkan/FamilyBudget/compare/0.1.1...0.2.0
 [diff: 0.1.0-0.1.1]: https://github.com/Luzkan/FamilyBudget/compare/0.1.0...0.1.1
 [diff: 0.0.0-0.1.0]: https://github.com/Luzkan/FamilyBudget/compare/0.0.0...0.1.0
+[0.4.3]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.4.3
 [0.4.2]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.4.2
 [0.4.1]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.4.1
 [0.4.0]: https://github.com/Luzkan/FamilyBudget/releases/tag/0.4.0